Moved changes over to main AreaDecomposition file. Removed AreaDecompositionDebugger.
